+++ tests/expand/generic_from_try_into_naked.rs
@@ -1,0 +1,12 @@
+disponent::declare!(
+    #[disponent::configure(from, try_into)]
+    pub enum Wrapper<T> {
+        Value(T),
+    }
+
+    pub trait Marker {}
+);
+
+fn main() {
+    let _wrapped: Wrapper<i32> = 5_i32.into();
+}